+++ src/main/java/nqueens/Main.java
@@ -2,7 +2,7 @@ package nqueens;
 
 public class Main {
     public static void main(String[] args) {
-	NaiveSolver ns = new NaiveSolver(8);
+	NaiveSolver ns = new NaiveSolver(8,false);
 	ns.printSolutions();
     }
 }